Modernise to Linux/wxWidgets/CMake; fix serial port and add a real TCP modem
GUI/build: replace the Qt5 front-end and qmake/Win32 build with a wxWidgets
(C++17) GUI and a CMake build system, Linux-only. Emulation runs on a worker
thread (EmulatorHost) talking to the UI via GuiBridge.

Serial: route the 16550 interrupt to IOMD IRQ B bit 2 (gated by MCR OUT2) to
match the RISC OS Serial driver, so transmit interrupts are delivered (was on
FIQ and never serviced). Replace the stub modem with a real telnet client -
ATDT dialling, IAC/binary negotiation for 8-bit-clean X/Y/ZMODEM, guard-timed
+++ escape, DTR/ATH hangup, non-blocking sockets. Expose a single Serial port.

Input: stop intercepting menu/function-key shortcuts so all keys (F1-F12, Ctrl
combos, Esc) reach RISC OS; only Ctrl+End releases the mouse / exits full-screen.

Cleanup/docs: drop Win32 and Qt5 sources and legacy cmos.ram/rpc.cfg; ignore
runtime/local state; rewrite README/COMPILE and add docs/peripherals.md.

+++ src/codegen_amd64.c
@@ -83,7 +83,7 @@ addrip(const void *addr)
 {
 	const ptrdiff_t rel = ((const char *) addr) -
 	                      ((const char *) &rcodeblock[blockpoint2][codeblockpos]);
-	addlong((uint32_t) (rel - 4));
+	addlong((uint32_t) (int32_t) (rel - 4));
 }
 
 static inline void
@@ -91,7 +91,7 @@ addrip_byte(const void *addr, uint8_t x)
 {
 	const ptrdiff_t rel = ((const char *) addr) -
 	                      ((const char *) &rcodeblock[blockpoint2][codeblockpos]);
-	addlong((uint32_t) (rel - 5));
+	addlong((uint32_t) (int32_t) (rel - 5));
 	addbyte(x);
 }
 
@@ -100,7 +100,7 @@ addrip_long(const void *addr, uint32_t x)
 {
 	const ptrdiff_t rel = ((const char *) addr) -
 	                      ((const char *) &rcodeblock[blockpoint2][codeblockpos]);
-	addlong((uint32_t) (rel - 8));
+	addlong((uint32_t) (int32_t) (rel - 8));
 	addlong(x);
 }
